# openapi.yaml
openapi: 3.0.3
info:
  title: OpenFairDB API
  version: 0.10.3
  contact:
    name: slowtec GmbH
    url: 'https://slowtec.de'
  license:
    name: AGPLv3
    url: 'https://github.com/slowtec/openfairdb/blob/master/LICENSE'
servers:
  - url: 'https://api.ofdb.io/v0/'
    description: Public production server
  - url: 'https://dev.ofdb.io/v0/'
    description: Public unstable development server
paths:
...
  '/entries':
    post:
      summary: Create an entry
      tags:
        - Entries/Places
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEntryWithLicense'
      security:
        - bearerAuth: []
        - captchaCookieAuth: []
        - jwtAuth: []
        - userEmailCookieAuth: []
      responses:
        '200':
          description: Successful response
  '/entries/{ids}':
    get:
      summary: Get multiple entries
      tags:
        - Entries/Places
      parameters:
        - $ref: '#/components/parameters/IdListPath'
        - $ref: '#/components/parameters/OrgTagFilter'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
  '/entries/{id}':
    put:
      summary: Update an entry
      description: |
        The edited entry must include the *next version* of this entry
        in the `version` field, where *next version* = *current version* + 1.
      tags:
        - Entries/Places
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEntryWithVersion'
      security:
        - bearerAuth: []
        - captchaCookieAuth: []
        - jwtAuth: []
        - userEmailCookieAuth: []
      responses:
        '200':
          description: Successful response
...
components:
  schemas:
    NewEntry:
      properties:
        title:
          type: string
        description:
          type: string
        lat:
          $ref: '#/components/schemas/Latitude'
        lng:
          $ref: '#/components/schemas/Longitude'
        street:
          type: string
        zip:
          type: string
        city:
          type: string
        country:
          type: string
        state:
          type: string
        contact_name:
          $ref: '#/components/schemas/ContactName'
        email:
          $ref: '#/components/schemas/ContactEmail'
        telephone:
          $ref: '#/components/schemas/ContactPhone'
        homepage:
          $ref: '#/components/schemas/Url'
        opening_hours:
          $ref: '#/components/schemas/OpeningHours'
        founded_on:
          $ref: '#/components/schemas/FoundingDate'
        categories:
          type: array
          items:
            type: string
        tags:
          $ref: '#/components/schemas/TagArray'
        image_url:
          $ref: '#/components/schemas/ImageUrl'
        image_link_url:
          $ref: '#/components/schemas/ImageLink'
        links:
          $ref: '#/components/schemas/CustomLinkList'
      required:
        - title
        - description
        - lat
        - lng
    NewEntryWithLicense:
      allOf:
        - $ref: '#/components/schemas/NewEntry'
        - type: object
          properties:
            license:
              $ref: '#/components/schemas/License'
          required:
            - license
    NewEntryWithVersion:
      allOf:
        - $ref: '#/components/schemas/NewEntry'
        - type: object
          properties:
            version:
              type: integer
    Entry:
      allOf:
        - $ref: '#/components/schemas/NewEntryWithLicense'
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/Id'
            version:
              type: integer
            created:
              type: integer
            ratings:
              type: array
              items:
                type: string
    ImageUrl:
      description: |
        The external URL for an image.
      allOf:
        - $ref: '#/components/schemas/Url'
    ImageLink:
      description: |
        A hyperlink behind an image that is supposed to be opened when
        clicking on the image.
      allOf:
        - $ref: '#/components/schemas/Url'
    CustomLink:
      description: |
        A custom hyperlink with an optional title and description.
      properties:
        url:
          $ref: '#/components/schemas/Url'
        title:
          type: string
        description:
          type: string
      required:
        - url
    CustomLinkList:
      description: |
        An unordered list of custom links.

        The ordering of list items may change between different
        revisions even if this list has not been modified. Don't
        make any assumptions about a particular ordering!
      type: array
      items:
        $ref: '#/components/schemas/CustomLink'
...
    Id:
      type: string
      minLength: 32
      maxLength: 32
      description: |
        Identifier of a resource
      example: 7cee99c287094a94acbdcf29ffff2e85
...
    IdList:
      type: string
      description: |
        Comma-separated list of identifiers
      example: 7cee99c287094a94acbdcf29ffff2e85,0884c4e86e404072b6874b99b7e32640,ed8a2aef20054102b20950b1b78eb581
...
    License:
      type: string
      minLength: 1
      example: ODbL-1.0
    Email:
      type: string
      minLength: 3
      description: |
        An e-mail address.
      example: john.smith@example.com
...
    ContactName:
      description: |
        The name of the contact person.
      allOf:
        - $ref: '#/components/schemas/PersonFullName'
    ContactEmail:
      description: |
        An e-mail address to get in contact.
      allOf:
        - $ref: '#/components/schemas/Email'
    PersonFullName:
      type: string
      description: |
        The full name of a person.
      example: John Smith
...
    Phone:
      type: string
      example: 001 123456789
    ContactPhone:
      description: |
        An phone number to get in contact.
      allOf:
        - $ref: '#/components/schemas/Phone'
    Tag:
      type: string
      minLength: 1
      example: non-profit
    TagArray:
      type: array
      items:
        $ref: '#/components/schemas/Tag'
      example: [organic, non-profit]
    Url:
      type: string
      minLength: 8
      example: https://www.slowtec.de/
...
    OpeningHours:
      type: string
      minLength: 4
      description: |
        The opening hours in OpenStreetMap format.

        Specification: https://wiki.openstreetmap.org/wiki/Key:opening_hours

        Reference implementations: https://github.com/opening-hours/

        Generator tool: https://projets.pavie.info/yohours/

        The service trims leading/trailing whitespaces and stores values as is.
        Values are currently not validated against the OSM format syntax!
      example: 24/7
...
    FoundingDate:
        description: |
          The date on which an organization, initiative, or company has been founded or established.
        type: string
        format: date
        example: 1945-10-24
...
    Latitude:
      type: number
      format: double
      minimum: -90.0
      maximum: 90.0
      example: 37.2
      description: Geographic latitude (in degrees)
    Longitude:
      type: number
      format: double
      minimum: -180.0
      maximum: 180.0
      example: 120.7
      description: Geographic longitude (in degrees)
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    IdListPath:
      name: ids
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IdList'
...
    OrgTagFilter:
      name: org_tag
      in: query
      required: false
      schema:
        type: string
      description: |
        A single hash tag that is moderated by an organization.

        Results will be filtered according to the current clearance status.
        Unclear revisions of entries will either be removed or replaced by older,
        already cleared revisions. Unclear revisions of entries are expected to
        only exist temporarily for a short time frame until a newer version has
        been cleared.

        If the provided hash tag is not moderated by any organization the results
        will not be filtered by clearance status.

        Only a single tag can be specified, because filtering according to
        clearance by multiple organizations could produce ambiguous and
        conflicting results.

        Only when used in search requests: The hash tag will be included in the
        query as a mandatory search criteria like any other tag. Omitting this
        parameter and instead appending the moderated hash tag as a regular tag
        to the `tags` parameter will also return unfiltered search results, i.e.
        current revisions of entries independent of their clearance status.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    captchaCookieAuth:
      type: apiKey
      in: cookie
      name: ofdb-captcha
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userEmailCookieAuth:
      type: apiKey
      in: cookie
      name: ofdb-user-email
...